requirements.txt
azure-iot-device>=2.0.0
urllib3<2.0.0
python-dotenv>=0.19.0 